gluon-mesh-babel: move code from handle_neighbour.c into neighbours-babel
After refactoring the respondd module to work with a separate thread,
handle_neighbour() is different in those two programs. A common
implemenation thus is impossible and the structure is removed.
