+++ package/gluon-mesh-babel/src/Makefile
@@ -27,5 +27,5 @@ LDFLAGS_JSONC = $(shell pkg-config --libs json-c)
 respondd.so: respondd.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) $(LDLIBS) -shared $(LDFLAGS_JSONC) -o $@ $^ -lgluonutil -lblobmsg_json -lubox -lubus -luci
 
-neighbours-babel: neighbours-babel.c handle_neighbour.c
+neighbours-babel: neighbours-babel.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) $(CFLAGS_JSONC) $(LDFLAGS) $(LDLIBS) $(LDFLAGS_JSONC) -o $@ $^
